+++ astroplan/__init__.py
@@ -14,8 +14,6 @@ from ._astropy_init import *
 if not _ASTROPY_SETUP_:
     from .utils import *
 
-    # Import everything in astroplan.core
     from .core import *
-
-    # Import everything in astroplan.exceptions
     from .exceptions import *
+    from .sites import *
