+++ astroplan/tests/test_core.py
@@ -4,14 +4,14 @@ from __future__ import (absolute_import, division, print_function,
 from astropy.coordinates import (EarthLocation, Latitude, Longitude, SkyCoord)
 import astropy.units as u
 from astropy.time import Time
-from astropy.tests.helper import remote_data, assert_quantity_allclose
+from astropy.tests.helper import remote_data
 import numpy as np
 from numpy.testing import assert_allclose
 import pytz
 import datetime
 import unittest
 
-from ..core import FixedTarget, Observer, get_site, get_site_names, add_site
+from ..core import FixedTarget, Observer
 from ..exceptions import TargetAlwaysUpWarning, TargetNeverUpWarning
 
 def test_Observer_constructor_location():
@@ -202,7 +202,6 @@ def print_pyephem_parallactic_angle():
     obs = ephem.Observer()
     obs.lat = '19:49:34.3848'
     obs.lon = '-155:28:19.1964'
-    obs.elevation = elevation.value
     obs.date = time.datetime
     pyephem_target1 = ephem.FixedBody()
     pyephem_target1._ra = ephem.degrees((LST - desired_HA_1).to(u.rad).value)
@@ -772,33 +771,6 @@ def test_timezone_convenience_methods():
     naive_dts = list(map(lambda t: t.replace(tzinfo=None), dts))
     assert all(naive_dts == times_dt_ndarray - datetime.timedelta(hours=4))
 
-def test_get_site():
-    # Compare to the IRAF observatory list available at:
-    # http://tdc-www.harvard.edu/iraf/rvsao/bcvcorr/obsdb.html
-    keck = get_site('keck')
-    lon, lat, el = keck.to_geodetic()
-    assert_quantity_allclose(lon, -1*Longitude('155:28.7', unit=u.deg),
-                             atol=0.001*u.deg)
-    assert_quantity_allclose(lat, Latitude('19:49.7', unit=u.deg),
-                             atol=0.001*u.deg)
-    assert_quantity_allclose(el, 4160*u.m, atol=1*u.m)
-
-    keck = get_site('ctio')
-    lon, lat, el = keck.to_geodetic()
-    assert_quantity_allclose(lon, -1*Longitude('70.815', unit=u.deg),
-                             atol=0.001*u.deg)
-    assert_quantity_allclose(lat, Latitude('-30.16527778', unit=u.deg),
-                             atol=0.001*u.deg)
-    assert_quantity_allclose(el, 2215*u.m, atol=1*u.m)
-
-def test_add_site():
-    # Test observatory can be added and retrieved
-    new_site_name = 'University of Washington'
-    new_site_location = EarthLocation(-122.3080*u.deg, 47.6550*u.deg, 0*u.m)
-    add_site(new_site_name, new_site_location)
-    retrieved_location = get_site(new_site_name)
-    assert retrieved_location == new_site_location
-
 class TestExceptions(unittest.TestCase):
     def test_rise_set_transit_which(self):
         lat = '00:00:00'
@@ -837,7 +809,3 @@ class TestExceptions(unittest.TestCase):
         with self.assertRaises(TypeError):
             obs = Observer(location=EarthLocation(0, 0, 0))
             obs.altaz(Time('2000-01-01 00:00:00'), ['00:00:00','00:00:00'])
-
-    def test_bad_site(self):
-        with self.assertRaises(KeyError):
-            get_site('nonexistent site')
